// frontend-app/src/app/core/models/schoolCard.model.ts
export interface SchoolCard {
    visible: boolean;
    weblink: string;
    ofstedLastInsp: string;
    rating: string;
    uniqueReferenceNumber: string;
    establishmentName: string;
    typeOfEstablishment: string;
    address: string;
    statutoryLowAge: string;
    statutoryHighAge: string;
    gender: string;
    religiousCharacter: string;
    location: object;
    image: string;
    telephoneNum:string;
    isAddedToCompare:boolean;
  }
  